Rudimentary AI and Stance Icons
Summary:
Added stance icons, ability to check for opponent selection, and
rudimentary AI.

Stance icons are currently colored circles. The sprite sheet is 50X150 and
each icon is 50X50. It should be the third UI source to be loaded in the
initialize function, so name accordingly.

Added a control in the AutomatedController to check for opponent's
selection, first checking for status of multiplayer, then calling the
appropriate function. Multiplayer is not implemented, so this currently
passes.

Added rudimentary AI. Currently returns a random command.